what about now

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -5,7 +5,7 @@
  *
  * Return: last digit of @n and print it
  */
-int print_last_digit(int n)
+int print_last_digit(long int n)
 {
 	if (n < 0)
 		n *= -1;
