+++ 0x02-functions_nested_loops/main.h
@@ -11,7 +11,7 @@ int _islower(int c);
 int _isalpha(int c);
 int print_sign(int c);
 int _abs(int);
-int print_last_digit(int);
+int print_last_digit(long int);
 void jack_bauer(void);
 void times_table(void);
 int add(int, int);
